～進捗状況～
①プレイヤーの移動制御について調整中

※明日の目標はギミックの完成と当たり判定の実装。

※プレイヤーの動き制御は一旦おいておいて、ゲームとしての機能完成に着手する。（どうすればいいのかなんとなく方向性が定まったため、レベルデザインも考慮して、残りはギミックの後に取り組むべき）

+++ ManagedDxlGame/program/game/[001]Common/[000]Object/[000]Stage/[003]Model/Model.h
@@ -23,7 +23,7 @@ public:
 		std::string s_texture_path;
 		int s_create_count_x;
 		int s_create_count_z;
-		int s_scale;
+		float s_scale;
 		float s_interval;
 		tnl::Vector3 s_offset;
 	};
